libsemanage: Fallback to semanage_copy_dir when rename() failed
In some circumstances, like semanage-store being on overlayfs, rename()
could fail with EXDEV - Invalid cross-device link. This is due to fact
that overlays doesn't support rename() is source and target are not on
the same layer, e.g. in rpm-ostree based containers. Even though it's
not atomic operation, it's better to try to to copy files from src to
dst on our own in this case. Next rebuild will probably not fail as the
new directories will be on the same layer.

Fixes: https://github.com/SELinuxProject/selinux/issues/343

Reported-by: Joseph Marrero Corchado <jmarrero@redhat.com>
Signed-off-by: Petr Lautrbach <plautrba@redhat.com>